% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_amf_credentials.R
\name{flux_amf_credentials}
\alias{flux_amf_credentials}
\title{Generate list of user info to pass to \code{flux_download()}}
\usage{
flux_amf_credentials(
  user_name = Sys.getenv("AMERIFLUX_USER_NAME", unset = NA_character_),
  user_email = Sys.getenv("AMERIFLUX_USER_EMAIL", unset = NA_character_),
  intended_use = Sys.getenv("AMERIFLUX_INTENDED_USE", unset = 6),
  description = Sys.getenv("AMERIFLUX_DESCRIPTION", unset = NA_character_)
)
}
\arguments{
\item{user_name}{Your AmeriFlux username.}

\item{user_email}{The email address associated with your AmeriFlux profile.}

\item{intended_use}{An integer 1--6 as follows:
\itemize{
\item 1 = Synthesis / network synthesis analysis
\item 2 = Land model/Earth system model
\item 3 = Remote sensing research
\item 4 = Other research
\item 5 = Education (Teacher or Student)
\item 6 = Other
}}

\item{description}{An optional description of the project.}
}
\value{
A list with elements \code{user_name}, \code{user_email}, \code{intended_use}, and
\code{description}.
}
\description{
A helper to generate a list of AmeriFlux user info to pass along to
\code{\link[=flux_download]{flux_download()}}.  Because these elements are by default pulled from
environment variables, it is recommended that you set them in a project-level
\code{.Renviron} file like \code{AMERIFLUX_USER_NAME=myusername}, etc.
}
