% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_badm.R
\name{flux_badm}
\alias{flux_badm}
\title{Read and tidy BADM subsets}
\usage{
flux_badm(
  manifest,
  variable_group,
  site_ids = NULL,
  networks = c("AMF", "CNF", "EUF", "FLX", "ICOS", "JPF", "KOF", "SAEON", "TERN")
)
}
\arguments{
\item{manifest}{A manifest data frame produced by \code{\link[=flux_discover_files]{flux_discover_files()}}.}

\item{variable_group}{A single \code{VARIABLE_GROUP} without a \code{GRP_} prefix.
Options can be viewed at
\url{https://ameriflux.lbl.gov/data/badm/badm-standards/}}

\item{site_ids}{A vector of site IDs to filter the manifest by.  If \code{NULL}
(the default), the manifest isn't filtered by site ID.}

\item{networks}{A character vector indicating which networks to extract files
from. Multiple values may be provided. Defaults to all networks.}
}
\value{
A tibble.
}
\description{
Reads in BADM data from "BIF" csv files, subsets to a single \code{VARIABLE_GROUP}
and returns a wide data frame.
}
\examples{
\dontrun{
manifest <- flux_discover_files()
flux_badm(manifest, "SOIL_CHEM")
flux_badm(manifest, "LAI")
}
}
