% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_duckdb.R
\name{flux_db_build}
\alias{flux_db_build}
\title{Ingest FLUXNET data into a DuckDB database}
\usage{
flux_db_build(
  manifest,
  duckdb_path = Sys.getenv("FLUXNET_DB_PATH", unset = "fluxnet/fluxnet.duckdb")
)
}
\arguments{
\item{manifest}{A file manifest tibble created by \code{\link[=flux_discover_files]{flux_discover_files()}}.}

\item{duckdb_path}{File path to a .duckdb file used to store FLUXNET data.
Can be set with the environment variable \code{FLUXNET_DB_PATH}.}
}
\value{
Returns nothing; called for side-effects only.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
As an alternative to reading FLUXNET data into memory with \code{\link[=flux_read]{flux_read()}},
this ingests all FLUXNET data into a DuckDB database that can be queried in R
with \code{dplyr} via \code{dbplyr}. This is especially useful for daily and hourly
data which are likely to overflow memory when trying to read them in with
\code{\link[=flux_read]{flux_read()}}.
}
\note{
Currently "BIF" and "BIFVARINFO" CSV files are \emph{not} included in the
DuckDB database.
}
\examples{
\dontrun{

# Ingest all available data into a DuckDB database
manifest <- flux_discover_files()
flux_db_build(manifest)
}

}
\seealso{
\code{\link[=flux_db_update]{flux_db_update()}}, \code{\link[=flux_db_connect]{flux_db_connect()}}
}
