% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_duckdb.R
\name{flux_db_connect}
\alias{flux_db_connect}
\title{Connect to (or create) DuckDB database for FLUXNET data}
\usage{
flux_db_connect(
  duckdb_path = Sys.getenv("FLUXNET_DB_PATH", unset = "fluxnet/fluxnet.duckdb")
)
}
\arguments{
\item{duckdb_path}{File path to a .duckdb file used to store FLUXNET data.
Can be set with the environment variable \code{FLUXNET_DB_PATH}.}
}
\value{
A database connection object of class \code{duckdb_connection}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
A convenience wrapper around \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.
}
\examples{
\dontrun{

con <- flux_db_connect()
DBI::dbListTables(con)
# [1] "annual"   "daily"    "hourly"   "manifest" "monthly"  "weekly"

# close connection
DBI::dbDisconnect(con)
}
}
\seealso{
\code{\link[=flux_db_build]{flux_db_build()}}
}
