% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_duckdb.R
\name{flux_db_update}
\alias{flux_db_update}
\title{Update existing DuckDB database with new FLUXNET data}
\usage{
flux_db_update(
  manifest,
  duckdb_path = Sys.getenv("FLUXNET_DB_PATH", unset = "fluxnet/fluxnet.duckdb")
)
}
\arguments{
\item{manifest}{A file manifest tibble created by \code{\link[=flux_discover_files]{flux_discover_files()}}.}

\item{duckdb_path}{File path to a .duckdb file used to store FLUXNET data.
Can be set with the environment variable \code{FLUXNET_DB_PATH}.}
}
\value{
Returns nothing; called for side-effects only.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Updates an existing DuckDB database crated by \code{\link[=flux_db_build]{flux_db_build()}}. When you've
downloaded and extracted new data since building the database, this may be
faster than rebuilding the database by deleting the .duckdb file and
re-running \code{\link[=flux_db_build]{flux_db_build()}}. If run successfully, it will update the
'manifest' table as well.
}
\note{
This will error when new CSVs contain columns not already in the
database. In this case, you'll have to delete the .duckdb file and run
\code{\link[=flux_db_build]{flux_db_build()}} again to regenerate the database.
}
\examples{
\dontrun{

# Build database
manifest <- flux_discover_files()
flux_db_build(manifest)

# Download and extract additional sites
flux_download()
flux_extract()

# Update database
manifest <- flux_discover_files()
flux_db_update(manifest)

}

}
\seealso{
\code{\link[=flux_db_build]{flux_db_build()}}, \code{\link[=flux_db_connect]{flux_db_connect()}}
}
