% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_discover_files.R
\name{flux_discover_files}
\alias{flux_discover_files}
\title{Create a "manifest" of downloaded and unzipped FLUXNET data files}
\usage{
flux_discover_files(
  data_dir = "fluxnet/unzipped",
  remove_duplicates = TRUE,
  ...
)
}
\arguments{
\item{data_dir}{The directory to look for FLUXNET CSV files in, typically the
same as the \code{output_dir} used for \code{\link[=flux_extract]{flux_extract()}}.}

\item{remove_duplicates}{Logical; if \code{TRUE} (default) duplicate rows will be
removed from the resulting dataset.  Note: this does \emph{not} remove the
duplicate files!}

\item{...}{Arguments passed to \code{\link[=flux_listall]{flux_listall()}}.}
}
\value{
Prints a summary of discovered available data and returns
(invisibly) a dataframe with file paths and metadata extracted from file
names and merged in from \code{\link[=flux_listall]{flux_listall()}}.
}
\description{
Create a "manifest" of downloaded and unzipped FLUXNET data files
}
\examples{
\dontrun{
# Download data
flux_download(site_ids = c("AU-Boy", "BR-CST"))

# Extract annual and monthly data
flux_extract(resolutions = c("y", "m"))

# Create a manifest of extracted files
manifest <- flux_discover_files()
}
}
