% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_listall.R
\name{flux_listall}
\alias{flux_listall}
\title{List available FLUXNET zip files for download}
\usage{
flux_listall(
  cache_dir = rappdirs::user_cache_dir("fluxnet"),
  cache_age = as.difftime(1, units = "days"),
  clean_cache = 10L,
  log_file = NULL,
  echo_cmd = FALSE,
  ...
)
}
\arguments{
\item{cache_dir}{The directory to store the list of available FLUXNET data
in.}

\item{cache_age}{A \code{difftime} object of length 1. If there are no cached
snapshots more recent than \code{cache_age}, a new one will be downloaded and
stored. You can force the cache to be invalidated with \code{cache_age = -Inf}.}

\item{clean_cache}{A number of files \eqn{\geq 1} to keep in \code{cache_dir}.
Defaults to 10, which keeps only the 10 most recent snapshots.}

\item{log_file}{An optional file path (e.g. \code{"log.txt"}) to direct the
\code{fluxnet-shuttle} log to. Useful for debugging.}

\item{echo_cmd}{Set to \code{TRUE} to print the shell command in the console.
Passed to \code{\link[processx:run]{processx::run()}}.}

\item{...}{Arguments passed to \code{\link[=flux_install_shuttle]{flux_install_shuttle()}}.}
}
\value{
A data frame of stations with available data and their metadata.
}
\description{
This provides a wrapper around the
\href{https://github.com/fluxnet/shuttle}{fluxnet-shuttle} command-line utility's
\code{listall} command, which downloads a data frame of available .zip files. By
default, the downloaded CSV is stored in
\code{rappdirs::user_cache_dir("fluxnet")}.  If there is allready a FLUXNET
shanpshot CSV file downloaded and it is more recent than \code{cache_age}, it will
be read in instead of downloading a new snapshot.
}
\note{
To force the \code{fluxnet} R package to re-install the \code{fluxnet-shuttle}
utility, remove the Pyhton virtualenv it is installed in by running
\code{reticulate::virtualenv_remove("fluxnet")}. Then, when you run
\code{flux_listall()} next, the virtualenv will be re-created and
\code{fluxnet-shuttle} will be re-installed.
}
\examples{
\dontrun{
fluxnet_files <- flux_listall()

# Invalidate cache and update it
fluxnet_files <- flux_listall(cache_age = -Inf)
}
}
