% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_map_sites.R
\name{flux_map_sites}
\alias{flux_map_sites}
\title{Map sites with downloaded data}
\usage{
flux_map_sites(
  manifest,
  color_var = c("data_hub", "igbp", "network", "first_year", "last_year")
)
}
\arguments{
\item{manifest}{A data manifest created by \code{\link[=flux_discover_files]{flux_discover_files()}}.}

\item{color_var}{A variable to use to color-code points.}
}
\value{
A ggplot2 object
}
\description{
Plot locations of sites in a file manifest on a world map.
}
\examples{
\dontrun{
manifest <- flux_discover_files()
flux_map_sites(manifest)
}

}
