% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_read.R
\name{flux_read}
\alias{flux_read}
\title{Read in FLUXNET data}
\usage{
flux_read(
  manifest,
  resolution = c("y", "m", "w", "d", "h"),
  datasets = c("ERA5", "FLUXMET"),
  networks = c("AMF", "CNF", "EUF", "FLX", "ICOS", "JPF", "KOF", "SAEON", "TERN"),
  site_ids = NULL
)
}
\arguments{
\item{manifest}{A manifest data frame produced by \code{\link[=flux_discover_files]{flux_discover_files()}}.}

\item{resolution}{The time resolution to read in.  Must be one of \code{"y"}
(annual), \code{"m"} (monthly), \code{"w"} (weekly), \code{"d"} (daily), or \code{"h"}
(hourly/half-hourly).}

\item{datasets}{Character vector of one or both of \code{"FLUXMET"} or \code{"ERA5"}.
Defaults to both.}

\item{networks}{A character vector indicating which networks to extract files
from. Multiple values may be provided. Defaults to all networks.}

\item{site_ids}{A vector of site IDs to filter the manifest by.  If \code{NULL}
(the default), the manifest isn't filtered by site ID.}
}
\description{
Reads and minimally cleans FLUXNET data found by \code{\link[=flux_discover_files]{flux_discover_files()}}.
}
\examples{
\dontrun{
manifest <- flux_discover_files()
daily <- flux_read(manifest, resolution = "d")
annual <- flux_read(manifest, resolution = "y")

# Filter manifest by metadata first
metadata <- flux_listall()

library(dplyr)
manifest_enriched <- left_join(manifest, metadata, by = join_by(site_id))
manifest_WET <- manifest_enriched \%>\% filter(igbp == "WET")
annual_wet <- flux_read(manifest_WET, resolution = "y")

}


}
