% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_fluxnet_zips.R
\name{unzip_fluxnet_zips}
\alias{unzip_fluxnet_zips}
\title{Unzip downloaded FLUXNET zip files}
\usage{
unzip_fluxnet_zips(
  location,
  exdir = file.path(location, "unzipped"),
  overwrite = TRUE
)
}
\arguments{
\item{location}{The path to the directory containing zip files.}

\item{exdir}{The directory to unzip to.}

\item{overwrite}{Logical; overwrite already unzipped files?}
}
\value{
A list of two data frames, one of successfully unzipped files and
one containing any failed files.
}
\description{
Unzips all CSV files in FLUXNET zip files
}
\examples{
\dontrun{
unzip_fluxnet_zips("fluxnet/icos/")
}


}
