% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_citations.R
\name{flux_citations}
\alias{flux_citations}
\title{Output required per-dataset citations for FLUXNET data}
\usage{
flux_citations(
  site_ids,
  output = c("data.frame", "bibtex"),
  output_path = NULL,
  ...
)
}
\arguments{
\item{site_ids}{Character vector of site IDs, e.g. \code{c("AR-Bal", "DE-Gwg")}.}

\item{output}{Either \code{"data.frame"} to return a tibble or \code{"bibtex"} to
return (or write) BibTeX entries. See '\strong{Value}' for more details.}

\item{...}{Additional arguments passed to \code{\link[=flux_listall]{flux_listall()}}.}

\item{bibtex_path}{Path to a .bib file to write BibTeX to, passed to the
\code{con} argument of \code{\link[=writeLines]{writeLines()}}. If \code{NULL} (default), BibTeX will be
returned as character. Has no effect if \code{output = 'data.frame'}.}
}
\value{
If \code{output = 'data.frame'}, a \code{tibble} is returned with a
\code{bibentry} list-column with elements of class \link{bibentry} and a \code{citation}
column with citations formatted in the default style (see the \code{format()}
method for \code{\link[=bibentry]{bibentry()}} for more details).  If \code{output = 'bibtex'}, BibTeX
entries are either returned as an atomic character vector (if \code{bibtex_path}
is \code{NULL}) or written to a file.
}
\description{
Given a vector of site IDs, this either returns a dataframe with or BibTeX
citations for each site.
}
\examples{

# Return dataframe with bibentries and formatted citations
flux_citations(c("AR-Bal", "DE-Gwg"))

# Return BibTeX
flux_citations(c("AR-Bal", "DE-Gwg"), output = "bibtex")

# Append BibTeX entries to a file
\dontrun{

flux_citations(
  c("AR-Bal", "DE-Gwg"),
  output = "bibtex",
  bibtex_path = "references.bib"
)

}

}
