% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_download.R
\name{flux_download}
\alias{flux_download}
\title{Download FLUXNET zip files}
\usage{
flux_download(
  file_list_df = NULL,
  site_ids = NULL,
  download_dir = "fluxnet",
  overwrite = FALSE,
  user_info = list(ameriflux = flux_amf_credentials()),
  ...
)
}
\arguments{
\item{file_list_df}{When \code{NULL} (default), \code{\link[=flux_listall]{flux_listall()}} is used to
determine the sites with data available to download (specific sites can be
selected with \code{site_ids}).  If \code{file_list_df} is supplied, it should be a
data frame generated by \code{\link[=flux_listall]{flux_listall()}}, but potentially filtered to
exlude some rows. This provides an alternative way of downloading only
specific sites. See the examples for a possible use case. If \code{file_list_df}
is not \code{NULL}, \code{cache_dir}, \code{use_cache}, and \code{cache_age} will be ingored
but \code{site_ids} will still be used.}

\item{site_ids}{If \code{NULL} (default) all available sites will be downloaded.
Alternatively, supply a character vector of site IDs. For example,
\code{c("UK-GaB", "CA-Ca2")}.}

\item{download_dir}{The directory to download zip files to.}

\item{overwrite}{Logical; overwrite already downloaded .zip files? If \code{FALSE}
it will skip downloading existing files, unless they are invalid .zip files
(e.g. due to partial download or corruption).}

\item{user_info}{An optional list with data-hub-specific user information.
Only AmeriFlux uses this currently. By default, these are retrieved from
environment variables by \code{\link[=flux_amf_credentials]{flux_amf_credentials()}}.}

\item{...}{Arguments passed to \code{\link[=flux_listall]{flux_listall()}}.}
}
\value{
Invisibly returns a tibble with the download URL, path on disk, HTTP
status code, and whether or not the download was successful.
}
\description{
Downloads zip files (one per site) for available FLUXNET sites.
}
\examples{
\dontrun{
# Download data for all available sites
flux_download()

# Download data for just select site IDs
flux_download(site_ids = c("UK-GaB", "CA-Ca2"))

# Download specific sites filtered by site metadata (IGPB  and data hub for example)
available_sites <- flux_listall()
to_get <-
  available_sites[available_sites$igbp == "CRO" & available_sites$data_hub == "AmeriFlux", ]
flux_download(file_list_df = to_get)

# Get a fresh list of available data and overwrite any existing downloads
flux_download(use_cache = FALSE, overwrite = FALSE)
}

}
