% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_download.R
\name{flux_download}
\alias{flux_download}
\title{Download FLUXNET zip files}
\usage{
flux_download(
  file_list_df = NULL,
  site_ids = "all",
  download_dir = "fluxnet",
  overwrite = FALSE,
  user_info = list(ameriflux = list(user_name = Sys.getenv("AMERIFLUX_USER_NAME"),
    user_email = Sys.getenv("AMERIFLUX_USER_EMAIL"), intended_use =
    Sys.getenv("AMERIFLUX_INTENDED_USE", unset = "6"), description =
    Sys.getenv("AMERIFLUX_DESCRIPTION"))),
  ...
)
}
\arguments{
\item{file_list_df}{When \code{NULL} (default), \code{\link[=flux_listall]{flux_listall()}} is used to
determine the sites with data available to download (specific sites can be
selected with \code{site_ids}).  If \code{file_list_df} is supplied, it should be a
data frame generated by \code{\link[=flux_listall]{flux_listall()}}, but potentially filtered to
exlude some rows. This provides an alternative way of downloading only
specific sites. See the examples for a possible use case. If \code{file_list_df}
is not \code{NULL}, \code{cache_dir}, \code{use_cache}, and \code{cache_age} will be ingored
but \code{site_ids} other than \code{"all"} will still be used.}

\item{site_ids}{Character; either \code{"all"} to download all sites available, or
a vector of site IDs. For example, \code{c("UK-GaB", "CA-Ca2")}.}

\item{download_dir}{The directory to download zip files to.}

\item{overwrite}{Logical; overwrite already downloaded .zip files? If \code{FALSE}
it will skip downloading existing files, unless they are invalid .zip files
(e.g. due to partial download or corruption).}

\item{user_info}{A list of optional information for specific data hubs.
Currently, only used by AmeriFlux.  For AmeriFlux, please include your
AmeriFlux \code{user_name}, \code{user_email}, \code{description}, and \code{intended_use}
(1--6) as follows:
\itemize{
\item 1 = Synthesis / network synthesis analysis
\item 2 = Land model/Earth system model
\item 3 = Remote sensing research
\item 4 = Other research
\item 5 = Education (Teacher or Student)
\item 6 = Other
}
It is recommended to set these options in a project-level \code{.Renviron} file,
as the default is to pull them from environment variables.}

\item{...}{Arguments passed to \code{\link[=flux_listall]{flux_listall()}}.}
}
\value{
Invisibly returns a tibble with the download URL, path on disk, HTTP
status code, and whether or not the download was successful.
}
\description{
Downloads zip files (one per site) for available FLUXNET sites.
}
\examples{
\dontrun{
# Download data for all available sites
flux_download()

# Download data for just select site IDs
flux_download(site_ids = c("UK-GaB", "CA-Ca2"))

# Download specific sites filtered by site metadata (IGPB  and data hub for example)
available_sites <- flux_listall()
to_get <-
  available_sites[available_sites$igbp == "CRO" & available_sites$data_hub == "AmeriFlux", ]
flux_download(file_list_df = to_get)

# Get a fresh list of available data and overwrite any existing downloads
flux_download(use_cache = FALSE, overwrite = FALSE)
}

}
