% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_download.R
\name{flux_download}
\alias{flux_download}
\title{Download FLUXNET zip files}
\usage{
flux_download(
  file_list_df = NULL,
  site_ids = "all",
  download_dir = "fluxnet",
  overwrite = FALSE,
  use_cache = TRUE,
  cache_dir = rappdirs::user_cache_dir("fluxnet"),
  cache_age = as.difftime(30, units = "days"),
  log_file = NULL,
  echo_cmd = FALSE
)
}
\arguments{
\item{file_list_df}{When \code{NULL} (default), \code{\link[=flux_listall]{flux_listall()}} is used to
determine the sites with data available to download (specific sites can be
selected with \code{site_ids}).  If \code{file_list_df} is supplied, it should be a
data frame generated by \code{\link[=flux_listall]{flux_listall()}}, but potentially filtered to
exlude some rows. This provides an alternative way of downloading only
specific sites. See the examples for a possible use case. If
\code{file_list_df} is not \code{NULL}, \code{cache_dir}, \code{cache}, and \code{cache_age} will be
ingored but \code{site_ids} other than \code{"all"} will still be used.}

\item{site_ids}{Character; either \code{"all"} to download all sites available, or
a vector of site IDs, for example \code{c("UK-GaB", "CA-Ca2")}.}

\item{download_dir}{The directory to download zip files to.}

\item{overwrite}{Logical; overwrite already downloaded .zip files?}

\item{use_cache}{Logical; use cached list of files available to download if
it exists and is not older than \code{cache_age}?}

\item{cache_dir}{The directory to store the list of available FLUXNET data
in.}

\item{cache_age}{A \code{difftime} object of length 1. If there are no cached
snapshots more recent than \code{cache_age}, a new one will be downloaded and
stored. You can force the cache to be invalidated with \code{cache_age = -Inf}.}

\item{log_file}{An optional file path (e.g. \code{"log.txt"}) to direct the
\code{fluxnet-shuttle} log to. Useful for debugging.}

\item{echo_cmd}{Set to \code{TRUE} to print the shell command in the console.
Passed to \code{\link[processx:run]{processx::run()}}.}
}
\value{
Invisibly returns the output of \code{\link[curl:multi_download]{curl::multi_download()}}, which
contains information on download success, download time, HTTP errors, etc.
}
\description{
Downloads zip files (one per site) for available FLUXNET sites.
}
\examples{
\dontrun{
# Download data for all available sites
flux_download()

# Download data for just select site IDs
flux_download(site_ids = c("UK-GaB", "CA-Ca2"))

# Download specific sites filtered by site metadata (IGPB  and data hub for example)
available_sites <- flux_listall()
to_get <-
  available_sites[available_sites$igbp == "CRO" & available_sites$data_hub == "AmeriFlux", ]
flux_download(file_list_df = to_get)

# Get a fresh list of available data and overwrite any existing downloads
flux_download(use_cache = FALSE, overwrite = FALSE)
}

}
