% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_extract.R
\name{flux_extract}
\alias{flux_extract}
\title{Extract FLUXNET data from downloaded zip files}
\usage{
flux_extract(
  zip_dir = "fluxnet",
  output_dir = fs::path(zip_dir, "unzipped"),
  site_ids = "all",
  resolutions = c("y", "m", "w", "d", "h"),
  extract_varinfo = TRUE,
  extract_txt = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{zip_dir}{The directory with the zip files}

\item{output_dir}{The directory to unzip files to.  Within this directory,
data files will be nested by site.}

\item{site_ids}{A character vector of site IDs (e.g. \code{c("AR-TF2", "CA-Ca2")})
can be supplied to only unzip data for certain sites.  If \code{"all"}
(default), all zip files found in zip_dir will be unzipped.}

\item{resolutions}{A character vector indicating which time resolutions to
extract.  Options are yearly (\code{"y"}), monthly (\code{"m"}), daily (\code{"d"}),
and hourly/half-hourly (\code{"h"}). Multiple options may be passed with all of
them as default.}

\item{extract_varinfo}{Logical; extract the BIF and BIFVARINFO files
containing variable information in the BADM Interchange Format? Defaults to
\code{TRUE}.}

\item{extract_txt}{Logical; extract the README.txt and
DATA_POLICY_LICENSE_AND_INSTRUCTIONS.txt files? Defaults to \code{FALSE}.}

\item{overwrite}{Logical; should existing extracted files be overwritten
(\code{TRUE}) or ignored (\code{FALSE})?}
}
\description{
Extracts data from zip files downloaded by \code{\link[=flux_download]{flux_download()}} with options to
extract only subsets of the files they contain.
}
