% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxnet_shuttle_executable.R
\name{flux_install_shuttle}
\alias{flux_install_shuttle}
\title{Install the fluxnet-shuttle CLI}
\usage{
flux_install_shuttle(
  venv = Sys.getenv("FLUXNET_VENV", unset = "fluxnet"),
  version = Sys.getenv("FLUXNET_SHUTTLE_VERSION", unset = "main"),
  from = c("github", "pypi"),
  reinitialize = FALSE
)
}
\arguments{
\item{venv}{A name to use for creating a virtual environment.  Defaults to
\code{"fluxnet"}, but we recommend using a project-specific virtual environment.
You can set this in a project-level \code{.Renviron} file as
\code{FLUXNET_VENV=myproject} and it will be pulled from there.}

\item{version}{A version tag (e.g. \code{"0.3.7"}) to install.  Defaults to GitHub
development version (for now). Can also be set as an environment variable
in \code{.Renviron}, E.g. \verb{FLUXNET_SHUTTLE_VERSION=0.3.6}}

\item{from}{Where to install from? Currently only \code{"github"} is available,
but eventually \code{"pypi"} will be an option to install from
\href{https://pypi.org}{PyPI}.}

\item{reinitialize}{Logical; if \code{TRUE}, the virtual environment specified in
\code{venv} will be removed and re-initialized.  Useful if you'd like to update
the version of \code{fluxnet-shuttle}.}
}
\value{
The path to the \code{fluxnet-shuttle} CLI executable, silently.
}
\description{
Uses \code{reticulate} to install the
\href{https://fluxnet.github.io/shuttle/}{fluxnet-shuttle} Python library and
command line interface into a virtual environment.  Required for
\code{\link[=flux_listall]{flux_listall()}} and \code{\link[=flux_download]{flux_download()}}.
}
\note{
This will be run automatically the first time you run \code{\link[=flux_listall]{flux_listall()}}
or \code{\link[=flux_download]{flux_download()}}, so it is not necessary to run this function
separately first.
}
\examples{
\dontrun{

# Standard install of development version
flux_install_shuttle()

# Specify a version
flux_install_shuttle(version = "0.3.5")

# When run a second time, even after restarting the R session, it skips
# installation as long as the `venv` exists unless `reinitialize = TRUE`
flux_install_shuttle(version = "0.3.6")

# If you want to update the version, set `reinitilaize = TRUE`
flux_install_shuttle(version = "0.3.6", reinitialize = TRUE)
}

}
