% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_listall.R
\name{flux_listall}
\alias{flux_listall}
\title{List available FLUXNET zip files for download}
\usage{
flux_listall(
  cache_dir = rappdirs::user_cache_dir("fluxnet"),
  use_cache = TRUE,
  cache_age = as.difftime(1, units = "days"),
  log_file = NULL,
  echo_cmd = FALSE
)
}
\arguments{
\item{cache_dir}{The directory to store the list of available FLUXNET data
in.}

\item{use_cache}{Logical; use cached list of files available to download if
it exists and is not older than \code{cache_age}?}

\item{cache_age}{A \code{difftime} object of length 1. If there are no cached
snapshots more recent than \code{cache_age}, a new one will be downloaded and
stored. You can force the cache to be invalidated with \code{cache_age = -Inf}.}

\item{log_file}{An optional file path (e.g. \code{"log.txt"}) to direct the
\code{fluxnet-shuttle} log to. Useful for debugging.}

\item{echo_cmd}{Set to \code{TRUE} to print the shell command in the console.
Passed to \code{\link[processx:run]{processx::run()}}.}
}
\value{
A data frame of stations with available data and their metadata.
}
\description{
This provides a wrapper around the
\href{https://github.com/fluxnet/shuttle}{fluxnet-shuttle} command-line utility's
\code{listall} command, which downloads a data frame of available .zip files. By
default, the downloaded CSV is stored in
\code{rappdirs::user_cache_dir("fluxnet")}.  If there is allready a FLUXNET
shanpshot CSV file downloaded and it is more recent than \code{cache_age}, it will
be read in instead of downloading a new snapshot (unless \code{use_cache = FALSE}).
}
\examples{
\dontrun{
fluxnet_files <- flux_listall()

# Ignore cache
fluxnet_files <- flux_listall(use_cache = FALSE)

# Invalidate cache and update it
fluxnet_files <- flux_listall(cache_age = -Inf)
}
}
