% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_qc.R
\name{flux_qc}
\alias{flux_qc}
\title{Flag overly-gapfilled data}
\usage{
flux_qc(data, qc_vars, max_gapfilled = 0.5, if_missing = c("flag", "ignore"))
}
\arguments{
\item{data}{A data frame created by \code{\link[=flux_read]{flux_read()}}.}

\item{qc_vars}{A character vector of column names with associated \verb{*_QC}
columns to use for flagging.}

\item{max_gapfilled}{Numeric between 0 and 1; cutoff for the \code{is_bad} flag to
be \code{TRUE}.}

\item{if_missing}{Set the behavior for rows with \code{NA} for the \verb{*_QC}
columns.  If \code{"flag"}, an \code{NA} will get the \code{is_bad = TRUE} flag.  If
\code{"ignore"}, the \code{is_bad} flag will be \code{NA} also.}
}
\value{
A tibble with the added columns \code{pct_gapfilled} (the maximum
proportion gapfilled of all the \code{qc_vars} for each row) and \code{is_bad} (a
flag indicating whether a row had \code{pct_gapfilled} > \code{max_gapfilled}).
}
\description{
Flags rows of data based on variables with associated \verb{_QC} columns.
}
\examples{
\dontrun{

# Flag rows where NEE_VUT_REF is more than 50\% gapfilled
manifest <- flux_discover_files()
annual <- flux_read(manifest, resolution = "y")
annual_flagged <- flux_qc(
  annual,
  qc_vars = "NEE_VUT_REF",
  max_gapfilled = 0.5,
  if_missing = "ignore"
)

# Treat `NA`s in the `NEE_VUT_REF_QC` column as over the threshold
annual_flagged2 <- flux_qc(
  annual,
  qc_vars = "NEE_VUT_REF",
  max_gapfilled = 0.5,
  if_missing = "flag"
)

}

}
