% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_varinfo.R
\name{flux_varinfo}
\alias{flux_varinfo}
\title{Read variable info from "BIFVARINFO" files}
\usage{
flux_varinfo(
  manifest,
  resolution = c("y", "m", "w", "d", "h"),
  networks = c("AMF", "CNF", "EUF", "FLX", "ICOS", "JPF", "KOF", "SAEON", "TERN"),
  site_ids = NULL
)
}
\arguments{
\item{manifest}{A manifest data frame produced by \code{\link[=flux_discover_files]{flux_discover_files()}}.}

\item{resolution}{The time resolution to read in.  Must be one of \code{"y"}
(annual), \code{"m"} (monthly), \code{"w"} (weekly), \code{"d"} (daily), or \code{"h"}
(hourly/half-hourly).}

\item{networks}{A character vector indicating which networks to extract files
from. Multiple values may be provided. Defaults to all networks.}

\item{site_ids}{A vector of site IDs to filter the manifest by.  If \code{NULL}
(the default), the manifest isn't filtered by site ID.}
}
\value{
A tibble
}
\description{
Extracts and tidies variable information from "BIFVARINFO" csv files.
}
\note{
This only returns variable info (\code{VARIABLE_GROUP == GRP_VAR_INFO}) from
the "BIFVARINFO" files as much of the other metadata they contain can be
found in the results of \code{flux_listall()}.

\code{HEIGHT} is returned as a character value because some heights are reported
as ranges and cannot be parsed as a single numeric value.
}
\examples{
\dontrun{
manifest <- flux_discover_files()
flux_varinfo(manifest)
}

}
