% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxnet-package.R
\docType{package}
\name{fluxnet-package}
\alias{fluxnet}
\alias{fluxnet-package}
\title{fluxnet: Download, Read, and Visualize FLUXNET Data}
\description{
Utility functions to help download, read in, and work with data from FLUXNET.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EcosystemEcologyLab/fluxnet-package}
  \item \url{https://ecosystemecologylab.github.io/fluxnet-package/}
  \item Report bugs at \url{https://github.com/EcosystemEcologyLab/fluxnet-package/issues}
}

}
\author{
\strong{Maintainer}: David J.P. Moore (\href{https://orcid.org/0000-0002-6462-3288}{ORCID})

Authors:
\itemize{
  \item David J.P. Moore (\href{https://orcid.org/0000-0002-6462-3288}{ORCID})
  \item Eric R. Scott \email{ericrscott@arizona.edu} (\href{https://orcid.org/0000-0002-7430-7879}{ORCID})
}

Other contributors:
\itemize{
  \item Arizona Board of Regents on behalf of The University of Arizona (\href{https://ror.org/0054f1w39}{ROR}) [copyright holder]
}

}
\keyword{internal}
